+++ tests/waveforms/test_waveform_representation.py
@@ -1,6 +1,7 @@
 """Tests for waveform representation."""
 
 from typing import override
+import sys
 import numpy as np
 from glyphtune import signal, waveforms
 
@@ -97,6 +98,32 @@ def test_operation_waveform_repr_with_operands_and_kwargs() -> None:
         f"{dummy_waveform_repr}, {dummy_waveform_repr}, kwarg=1)"
     )
     assert repr(nonsensical_waveform) == reference_repr
+
+
+def test_resample_waveform_repr() -> None:
+    """Ensure the representation of the waveform is as expected."""
+    sig = signal.Signal([[1, 2, 3], [2, 3, 4]])
+    wave = waveforms.ResampleWaveform(sig, 10)
+
+    assert repr(wave) == f"ResampleWaveform({repr(sig)}, 10)"
+
+
+def test_resample_waveform_repr_with_time_multiplier_argument() -> None:
+    """Ensure the representation of the waveform is as expected."""
+    sig = signal.Signal([[1, 2, 3], [2, 3, 4]])
+    wave = waveforms.ResampleWaveform(sig, 10, -1)
+
+    assert repr(wave) == f"ResampleWaveform({repr(sig)}, 10, time_multiplier=-1)"
+
+
+def test_resample_waveform_repr_long_array_full() -> None:
+    """Ensure the representation of the waveform is as expected."""
+    sig = signal.Signal(np.zeros((1, 1001)))
+    wave = waveforms.ResampleWaveform(sig, 10)
+
+    with np.printoptions(threshold=sys.maxsize):
+        sig_full_repr = repr(sig)
+    assert wave.full_repr() == f"ResampleWaveform({sig_full_repr}, 10)"
 
 
 def test_periodic_wave_repr() -> None: